useWeatherを分離
