+++ src/hooks/useWeather.ts
@@ -1,60 +1,7 @@
 import { useState, useEffect } from "react";
-import { useCurrentPosition, Position } from "./useCurrentPosition";
-import {
-  OPEN_WEATHER_ENDPOINT_GEOCORDING,
-  OPEN_WEATHER_ENDPOINT_ONECALL,
-} from "../constans";
-import { WeatherMap } from "../component/weatherCard";
-
-type Hour = {
-  "1h": number;
-};
-
-export type Weather = {
-  description: string;
-  icon: string;
-  id: number;
-  main: WeatherMap;
-};
-
-type dailyTemp = {
-  morn: number;
-  day: number;
-  eve: number;
-  night: number;
-  min: number;
-  max: number;
-};
-
-export type WeatherInfo = {
-  clouds: number;
-  dew_point: number;
-  dt: number;
-  feels_like: number;
-  humidity: number;
-  pressure: number;
-  sunrise: number;
-  sunset: number;
-  temp: number | dailyTemp;
-  uvi: number;
-  visibility: number;
-  weather: Weather[];
-  wind_deg: number;
-  wind_speed: number;
-  wind_gust?: number;
-  rain?: Hour;
-  snow?: Hour;
-};
-
-type ResponseWeather = {
-  current: WeatherInfo;
-  daily: WeatherInfo[];
-  hourly: WeatherInfo[];
-  lat: number;
-  lon: number;
-  timezone: string;
-  timezone_offset: number;
-};
+import { useCurrentPosition } from "./useCurrentPosition";
+import { getWeather, ResponseWeather } from "../lib/getWeather";
+import { getCityName } from "../lib/getCityName";
 
 export const useWeather = () => {
   const [isLoading, setIsLoading] = useState(false);
@@ -63,33 +10,6 @@ export const useWeather = () => {
   const [position, getPosition] = useCurrentPosition();
 
   const reload = () => getPosition();
-
-  const getCityName = async (position: Position): Promise<string> => {
-    const res = await fetch(
-      `${OPEN_WEATHER_ENDPOINT_GEOCORDING}?lat=${position.lat}&lon=${
-        position.lon
-      }&limit=1&lang=ja&appid=${import.meta.env.VITE_API_KEY}`
-    );
-
-    const city = await res.json();
-
-    return city.pop().local_names.ja;
-  };
-
-  const getWeather = async (position: Position): Promise<ResponseWeather> => {
-    const res = await fetch(
-      `${OPEN_WEATHER_ENDPOINT_ONECALL}?lat=${position.lat}&lon=${
-        position.lon
-      }&exclude=minutely,alerts&units=metric&lang=ja&appid=${
-        import.meta.env.VITE_API_KEY
-      }
-      `
-    );
-
-    const weather = await res.json();
-
-    return weather;
-  };
 
   useEffect(() => {
     setIsLoading(false);
